Create initial redaction pipeline
Adds a script, `imagedephi.py` to use as a CLI for redacting whole slide images.
As a first pass, this simply takes a tiff image file and overwrites its
ImageDescription metadata properties to read "Redacted by Image DePHI." This is
meant to be a starting point for discussion on how to build out a larger, more
robust system for WSI redaction.