Make rule.replace_value None when appropriate
If redact method is NOT replace, the rule.replace_value should be None, instead of
an arbitrary value like empty string. This does anger the type checker for the
`apply` method, so an extra check has been added to make sure that rule.replace_value
is NOT None before attempting to replace the value. I'm open to reverting this change
or handling that possibility in a different way if it adds any confusion.