WIP Use new pydantic models
The `RedactionPlan` objects now use the pydantic models for
rulesets (Ruleset) and rules (ConcreteMetadataRule). Some of
the behavior that used to be defined in rule classes was moved
to the plan class. This includes printing the rule in the report_plan
functions. Additionally, since we're now matching rules by tag name
instead of tag value, we need to also check `altnames`.

This commit also functionally turns off redaction, so it should be considered
a WIP.