Remove `TiffBasedMetadataRedactionPlan`
Now, there is a standalone function that can be used to build redation
plans of theoretically any type. This involved moving the save
functionality to the `RedactionPlan` subclasses. Additionally, the
subclasses no longer have to extract the necessary rules in their
constructor, and the build function takes care of that.

Co-authored-by: Brian Helba <brianhelba@users.noreply.github.com>