Add `--profile` flag to CLI
This allows users to specify which base rule set to use. It also removes
the existing flag `--strict` in favor of using `--profile strict`. This
is a breaking change, as calls to the CLI using `--strict` will no
longer work. This was done to avoid enabling users from specifying
multiple redacion formats at once.