Add flexibility for redacting custom DICOM tags
Replaces the `delete_custom_metadata` option in the rule set with
`custom_metadata_action`. This can either be omitted, in which case a
rule is needed for each piece of custom metadata in order for the image
to be redacted, or set to `"keep"` or `"delete"`. If an action is
specified, it will be applied if there is no rule for the custom
metadata element.